+++ packages/reactivity/__tests__/reactive.spec.ts
@@ -1,7 +1,7 @@
-import { isRef, ref } from '../src/ref-old'
+import { isRef, ref } from '../src/ref'
 import { isReactive, markRaw, reactive, toRaw } from '../src/reactive'
-import { computed } from '../src/computed-old'
-import { effect } from '../src/effect-old'
+import { computed } from '../src/computed'
+import { effect } from '../src/effect'
 
 describe('reactivity/reactive', () => {
   test('Object', () => {
