+++ packages/reactivity/src/dep-old.ts
@@ -1,0 +1,17 @@
+import type { ReactiveEffect } from './effect-old'
+import type { ComputedRefImpl } from './computed-old'
+
+export type Dep = Map<ReactiveEffect, number> & {
+  cleanup: () => void
+  computed?: ComputedRefImpl<any>
+}
+
+export const createDep = (
+  cleanup: () => void,
+  computed?: ComputedRefImpl<any>,
+): Dep => {
+  const dep = new Map() as Dep
+  dep.cleanup = cleanup
+  dep.computed = computed
+  return dep
+}
